define the key of the map returned by call_trees to Cow<str>
The reason is that the android struct implementation will return a map
with int as a key.
This will be fine for old legacy profile, but for new continuous
profile, the AndroidChunk implementation needs to return a map with a
string so we'll have to convert on the fly int to str in a wrapper
method. Given that in that case we won't have a ref to a tid string,
for android we'll have to create an owned str, whereas for the
SampleChunk, given it already have the tid defined as str we can
efficiently reuse those.